Remove CsvHeaderProcessorFactories, rename AltoidaCsvFilePreProcessor to AltoidaSummaryCsvPreProcessor, and fix AltoidaSummaryCsvPreProcessor
